+++ frontend-sveltekit/src/lib/components/form/formError.svelte
@@ -1,5 +1,18 @@
+<script lang="ts" context="module">
+	import { writable } from 'svelte/store';
+
+	export const formError = writable<string>(null);
+
+	export function setFormError(message: string) {
+		formError.set(message);
+	}
+
+	export function clearFormError() {
+		formError.set(null);
+	}
+</script>
+
 <script lang="ts">
-	import { formError } from './formErrorStore';
 	import { icons } from './icons';
 
 	export let closeButton: boolean = true;
