+++ frontend-sveltekit/src/lib/components/form/index.ts
@@ -1,10 +1,10 @@
 import Form from './form.svelte';
-import FormError from './formError.svelte';
 import TextField from './textField.svelte';
 import TextAreaField from './textAreaField.svelte';
 import RadioField from './radioField.svelte';
 import SubmitButton from './submitButton.svelte';
-import { setFormError } from './formErrorStore';
+import FormError from './formError.svelte';
+import { setFormError } from './formError.svelte';
 
 export {
 	Form,
